backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
transformers
torch
python-dotenv
gunicorn
